add alternate channel bonding mechanism
Original channel bonding mechanism would reset channel alignment
every time there was a symbol error in a control character, leading
to significant instability in image acquisition on links with
poor quality.

The new channel bonding mechanism picks an alignment and sticks
with it (in part because the auto-phase aligner engine also picks
an alignment and sticks with it). This leads to better link
stability.